+++ src/views/Playgrounds/index.ts
@@ -1,0 +1,1 @@
+export { default as PlaygroundForm } from './PlaygroundForm'
